[imgnormalize/data] New plots + bugfixes
I have run all the tests again to see if there were some glitches that
caused the strange behavior in the earth-huge test. The results are
actually the same as before, so we should investigate the code to see
why in the earth-huge case MPI is consistently (slightly) better than
OpenMP.

Anyway, I substituted all the data and the plots with the new ones.